LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.3_Line_Terminators/S7.3_A3.2_T3-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Expected an identifier but found 'comment' instead
S7.3_A3.2_T3

PASS Expected parsing failure

TEST COMPLETE
